Make LineCurrentFreeSpace the same as LineCurrentWholeSpace
